requirements.txt
numpy==1.16.1
grpcio==1.28.1
grpcio-tools==1.28.1
